requirements.txt
discord.py
python-dotenv
hvac
